Rename Validator backend confusable route to api route
